+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: support `Configuration` property type with `configurationTemplate` reference
+* `FEAT`: support embedding configuration templates via `configurationTemplates`
+
 ## 0.42.0
 
 * `FEAT`: support `isEmpty` condition ([#252](https://github.com/camunda/element-templates-json-schema/pull/252))